Load PREDICT-2 EIDITH data from the local database

These functions load data from the locally-stored SQLite database of downloaded and cleaned EIDITH data. ed_table is a general function, and ed_table_ it's standard evaluation equivalent, useful for programming. The other functions are convenience aliases for the individual tables. Functions beginning with the ed prefix refer to PREDICT-1 data tables, while ed2 functions refer to PREDICT-2 data tables.

ed2_events(...)

ed2_animals(...)

ed2_specimens(...)

ed2_animal_production(...)

ed2_crop_production(...)

ed2_zoo_sanctuary(...)

ed2_wildlife_restaurant(...)

ed2_natural_areas(...)

ed2_market_value_chain(...)

ed2_human(...)

ed2_tests(...)

ed2_test_serology(...)

ed2_test_interpreted(...)

ed2_human_animal_production(...)

ed2_extractive_industry(...)

ed2_dwellings(...)

ed2_human_crop_production(...)

ed2_human_extractive_industry(...)

ed2_human_hospital_worker(...)

ed2_human_hunter(...)

ed2_human_market(...)

ed2_human_restaurant(...)

ed2_human_sick_person(...)

ed2_human_temporary_settlements(...)

ed2_human_zoo(...)

ed2_behavior(...)

ed2_training(...)

Arguments

...
arguments passed to dplyr::filter() to subset data

Value

a tibble-style data frame.

Details

These functions take dplyr::filter() arguments to sub-set the data. Using these, the data is subsetted via SQL before it is loaded into memory. For large tables, such as the tests table, this is useful for reducing the memory footprint of your R session.

Note that subsetting in SQL is more limited:

Use 0 or 1 instead of TRUE or FALSE

Dates are stored as character strings, but as they are in YYYY-MM-DD format, filtering such as event_date > "2014-01-01" still works.