Export FASTA files for viral interpretation

ed_fasta() and ed_fasta_group() convert a data frame to a FASTA file. ed_tests_report() is a standard report returning a data frame of all uninterpreted test sequences, along with a set of metadata, which is commonly printed to FASTA. ed_report_excel() is a convenience function for creating excel-ready CSVs for reporting. See the FASTA vignette for more guidance.

ed_fasta(.data, filename = "", sequence = "sequence", ...)

ed_fasta_(.data, filename = "", sequence = "sequence", ..., .dots)

ed_tests_report(status = c("Result finalized, interpretation pending"),
  test_ids = NULL, meta_fields = c(sp_idname = "specimen_id_name", sp =
  "species_scientific_name", country = "country", an_idname =
  "animal_id_name", lab = "diag_lab_shortname", test_rq = "test_requested",
  meth = "test_requested_protocol", seq_id = "sequence_id", test_id =
  "test_id"))

ed_fasta_group(.data, filepath = "", sequence = "sequence",
  grouping = "virus", ...)

ed_report_excel(.data, filename)

Arguments

.data
The data frame

filename
name of file to write to. If "" (default), the output is written to console.

sequence
the column name in the data frame that contains the sequence. Default is "sequence".

...
Additional columns to include as metadata, using dplyr::select() syntax. If nothing is included, all columns are used.

.dots
Columns specified using standard evaluation.

status
For ed_test_report(), the status of tests that should be pulled. A vector of any of "Interpretation completed", "Active testing ongoing - pooled test positive", "Pooled test positive", "Active testing ongoing", or "Under Sequence Review". A warning will be printed if some tests have no sequence data.

test_ids
For ed_test_report(). If provided, status is ignored and tests are pulled by numeric test_id values

meta_fields
For ed_test_report(). A vector of names of metadata to provide. If named, fields will be renamed, which is useful for abbreviating fields for FASTA export

filepath
File path indicating where grouped FASTA files should be saved.

grouping
The variable used to group sequences into separate FASTA files. One of "virus", "method", or "both".

Value

For ed_fasta(), ed_fasta_group(), and ed_report_excel(), the file(s) will be written to a file or the console (default). Also, a character vector of each FASTA record is returned invisibly ed_tests_report() returns a data frame.

Examples

# NOT RUN {
 ed_fasta(ed_tests_report())

# }