Functions to download EIDITH tables via API

These functions download data directly from the EIDITH API. They require authorization. They can be useful for comparing local data against data updated in the database, or processed vs. unprocessed data. The table functions load data from the local database instead, and are thus faster and work without an internet connection.

ed_get(endpoint, verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, auth = NULL, ...)

ed_get_events(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, ...)

ed_get_animals(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, ...)

ed_get_specimens(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, ...)

ed_get_tests(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, ...)

ed_get_viruses(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, ...)

ed_get_testspecimen(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01",
  lmdate_to = Sys.Date() + 1, ...)

Arguments

verbose
Show a progress bar and other messages?

postprocess
Should data be cleaned via ed_process() or returned raw?

header_only
Return only the table header. Useful for checking if API access works

lmdate_from
filter records by earliest date, in YYYY-MM-DD format

lmdate_to
filter records by latest date, in YYYY-MM-DD format (see details)

...
additional arguments passed to httr::GET()

Value

a tibble-style data frame