Load mock EIDITH data for PREDICT-1 tutorials
Source: R/mock_data_calls.R

These functions load mock versions of the EIDITH PREDICT-1 data tables to be used for demonstrations, tutorials, and practice. They are drop-in replacements for the ed_table() functions.

ed_table_mock_(mock_table, ..., .dots, raw = FALSE)

ed_table_mock(mock_table, ...)

ed_events_mock(...)

ed_animals_mock(...)

ed_specimens_mock(...)

ed_tests_mock(...)

ed_viruses_mock(...)

ed_testspecimen_mock(...)

Arguments

mock_table
one of the EIDITH database tables. One of "events", "animals", "specimens", "tests", "viruses", or "test_specimen_ids".

...
arguments passed to dplyr::filter() to subset data

.dots
standard-evaluation versions of subsetting arguments

raw
Whether to return the unprocessed form of the mock data. This is equivalent to calling the ed_get() functions with postprocess=FALSE

Value

a tibble-style data frame.