Connecting eidith tables

To understand how various eidith tables are connected via their column names. Visualizing these connections is helpful in determining how to join eidith tables for further analyses.

ed_tables_conn(cols = "id")

Arguments

cols

can be either all or id. If all, all common columns are presented. If id, only unique id columns are shown. Default value of type is id.

Value

a data frame with three columns: T1 and T2 corresponding to eidith tables, and conn, the connecting column.

Examples

ed_tables_conn()

#>         T1               T2        conn
#> 1    Event           Animal    event_id
#> 2   Animal         Specimen   animal_id
#> 3 Specimen TestIDSpecimenID specimen_id
#> 4     Test            Virus     test_id
#> 5     Test TestIDSpecimenID     test_id
#> 6    Virus TestIDSpecimenID     test_id

ed_tables_conn(cols = "all")

#>          T1               T2                conn
#> 1     Event           Animal       database_date
#> 2     Event           Animal        date_created
#> 3     Event           Animal       date_modified
#> 4     Event           Animal            event_id
#> 5     Event         Specimen       database_date
#> 6     Event         Specimen        date_created
#> 7     Event         Specimen       date_modified
#> 8     Event             Test       database_date
#> 9     Event             Test        date_created
#> 10    Event             Test       date_modified
#> 11    Event            Virus       database_date
#> 12    Event            Virus        date_created
#> 13    Event            Virus       date_modified
#> 14   Animal         Specimen           animal_id
#> 15   Animal         Specimen       database_date
#> 16   Animal         Specimen        date_created
#> 17   Animal         Specimen       date_modified
#> 18   Animal             Test       database_date
#> 19   Animal             Test        date_created
#> 20   Animal             Test       date_modified
#> 21   Animal            Virus       database_date
#> 22   Animal            Virus        date_created
#> 23   Animal            Virus       date_modified
#> 24 Specimen             Test       database_date
#> 25 Specimen             Test        date_created
#> 26 Specimen             Test       date_modified
#> 27 Specimen             Test       specimen_type
#> 28 Specimen             Test specimen_type_group
#> 29 Specimen            Virus       database_date
#> 30 Specimen            Virus        date_created
#> 31 Specimen            Virus       date_modified
#> 32 Specimen TestIDSpecimenID         specimen_id
#> 33     Test            Virus       database_date
#> 34     Test            Virus        date_created
#> 35     Test            Virus       date_modified
#> 36     Test            Virus      interpretation
#> 37     Test            Virus            sequence
#> 38     Test            Virus             test_id
#> 39     Test            Virus          virus_name
#> 40     Test TestIDSpecimenID             test_id
#> 41    Virus TestIDSpecimenID             test_id