eidith

The eidith R package provides programmatic access and analytical tools for data from the PREDICT program, housed at the Emerging Infectious Disease Information Technology Hub.

Details

The eidith package contains no data. To access data, you must be registered EIDITH user with data access privileges.

Topics

Authentication with EIDITH credentials

Downloading EIDITH data to a local store and direct API access

Data cleaning

Metadata for tables and fields in the EIDITH database

Contact

Package source code is found at https://github.com/ecohealthalliance/eidith

File bug reports and request features at https://github.com/ecohealthalliance/eidith/issues

Online version of this documentation found at https://ecohealthalliance.github.io/eidith/

Questions related to your EIDITH credentials should be directed to technology@eidith.org.

To learn more about eidith, start with the vignettes: browseVignettes(package = "eidith")