EIDITH logon credentials

These are convience functions for using your default EIDITH logon credentials for any calls to EIDITH. Place your EIDITH username and password in your home .Renviron file as EIDITH_USERNAME and EIDITH_PASSWORD and these values will be used as defaults for all calls.

eidith_user(verbose = interactive())

eidith_pwd(verbose = interactive())

Arguments

verbose

Show messages?