Download EIDITH data to local storage

This function fetches all available tables from the EIDITH database and caches the data locally for further use with the table functions. The database persists between R sessions, so you only need to run this when you want to update data. You should also run it when updating to new versions of the eidith package.

ed_db_download(verbose = interactive())

Arguments

verbose

Show messages while in progress?

Details

This function does a few things:

Fetches all tables from EIDITH (events, animals, specimens, tests, viruses, and test-specimen cross-references)

Cleans and pre-processes the data with ed_process().

Stores the data in an SQLite database in your user data directory.

Each time it is run it clears out the previous data and downloads all the data, as such it takes a few minutes. In the future, we will allow for updating only with new and changed records.

The function will prompt for username and password unless you have cached your credentials.

See also

ed_db_status(), ed_db_updates(), ed_db_export()