Connections between tables

To understand how various EIDITH tables are connected via their column names. Visualizing these connections is helpful in determining how to join eidith tables for further analyses, and is available for both PREDICT 1 ed_tables_conn and PREDICT 2 ed2_tables_conn databases.

ed_tables_conn(cols = "id")

ed2_tables_conn(cols = "id")

Arguments

cols
can be either all or id. If all, all common columns are presented. If id, only unique id columns are shown. Default value is id.

Value

a data frame with three columns: T1 and T2 corresponding to eidith tables, and conn, the connecting column.

Examples

ed_tables_conn()

#>                 T1               T2        conn
#> 1           Animal         Specimen   animal_id
#> 2           Animal            Event    event_id
#> 3 TestIDSpecimenID             Test     test_id
#> 4 TestIDSpecimenID            Virus     test_id
#> 5         Specimen TestIDSpecimenID specimen_id
#> 6             Test            Virus     test_id

ed_tables_conn(cols = "all")

#>                 T1               T2        conn
#> 1           Animal         Specimen   animal_id
#> 2           Animal            Event    event_id
#> 3 TestIDSpecimenID             Test     test_id
#> 4 TestIDSpecimenID            Virus     test_id
#> 5         Specimen TestIDSpecimenID specimen_id
#> 6             Test            Virus     test_id

ed2_tables_conn()

#> # A tibble: 25 x 3
#>    T1    T2                 conn      
#>    <chr> <chr>              <chr>     
#>  1 Event ZooSanctuary       event_name
#>  2 Event AnimalProduction   event_name
#>  3 Event CropProduction     event_name
#>  4 Event Dwellings          event_name
#>  5 Event ExtractiveIndustry event_name
#>  6 Event WildlifeRestaurant event_name
#>  7 Event NaturalAreas       event_name
#>  8 Event MarketValueChain   event_name
#>  9 Event Animal             event_name
#> 10 Event Human              event_name
#> # ... with 15 more rows