eidith

The eidith R package provides programmatic access and analytical tools for data from the http://www.vetmed.ucdavis.edu/ohi/predict/, housed at the https://www.eidith.org/.

Details

The eidith package contains no data. To access data, you must be https://www.eidith.org/register.aspx EIDITH user with data access privileges.

Topics

Authentication with EIDITH credentials

Downloading EIDITH data to a local store and direct API access

Data cleaning

Metadata for tables and fields in the EIDITH database

Contact

Package source code is found at https://github.com/ecohealthalliance/eidith

File bug reports and request features at https://github.com/ecohealthalliance/eidith/issues

Online version of this documentation found at https://ecohealthalliance.github.io/eidith/

Questions related to your EIDITH credentials should be directed to mailto:technology@eidith.org.